using pt1FilterApplyWeighted for fyro lpf

+++ src/main/sensors/gyro.c
@@ -632,9 +632,8 @@ void dynLpfGyroUpdate(float throttle)
 
         float weight = 1.0f;
         if (gyro.dynLpfThrPercent) {
-            const float throttlePercent = gyro.dynLpfThrPercent / 100.0f;
-            //weight = MIN(throttle / throttlePercent, 1.0f);
-            const float minWeight = gyro.dynLpfMinWeight / 100.0f;
+            const float throttlePercent = (float)gyro.dynLpfThrPercent / 100.0f;
+            const float minWeight = (float)gyro.dynLpfMinWeight / 100.0f;
             weight = MIN(scaleRangef(throttle, 0.0f, throttlePercent, minWeight, 1.0f), 1.0f);
         }
